change HealthPool ints to doubles

+++ src/game/entity/HealthPool.h
@@ -5,17 +5,17 @@
 #include <SFML/System/Vector2.hpp>
 class HealthPool {
     public:
-    explicit HealthPool(int max_hp);
-    void takeDamage(int damageAmount);
+    explicit HealthPool(double max_hp);
+    void takeDamage(double damageAmount);
     bool isDead();
 
     void draw(sf::Vector2f display_position);
 
     private:
     sf::RectangleShape m_rect;
-    int m_max_health;
-    int m_current_health;
-    int m_block;
+    double m_max_health;
+    double m_current_health;
+    double m_block;
 };
 
 #endif
